add E2E tests
