+++ components/alerts/AlertsIntroPanel.vue
@@ -16,7 +16,7 @@ const emit = defineEmits(["dateRangeChanged"]);
 </script>
 
 <template>
-  <div class="space-y-4">
+  <div class="space-y-4" data-testid="alerts-intro-panel">
     <div class="rounded-lg border bg-card text-card-foreground shadow-sm">
       <div class="p-6 space-y-4">
         <img
